+++ NPCMapLocations/docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming version
 * Improved the new `npc_map_locations_summary` command:
+  * Added a list of unknown locations.
   * It now shows built-in map vectors too.
 * Fixed farm buildings not shown on load if any location without map vectors is also excluded.